requirements.txt
openai
gitpython
nltk